0.1.3 Alternate Myrmex Follow Range, Reset Task Tweaks
Follow range is now 50 by default instead of 128
To help stranded myrmex find their way home with the new changes, pathfinding may move in the general direction of the hive instead of to it.
Does have the side effect of possibly more pathfinding load, but shouldn't be as bad as having the chunks stay loaded.

Wrapped MoveThroughHive, and removed the path removal on Reset Task for it and LeaveHive, should help avoid stuttering sentinels

Queen doesn't use the alternate navigator anymore, queen hardly moves to justify the performance hit.